Reworked mid exam task.
1. Renamed classes, due to the inconsistency of the names with what they are appropriate for
- GreenSquares -> MidExam
- GreenSquaresSerive -> removed
- GreenSquaresConstants -> MidExamConstants
- SuperWindowProgrom-> removed

2. Changes in classes
2.1. MidExam
The main class with logic for drawing objects and processing user movements

2.2. GreenSquaresServe
The class was seen because it did not at all correspond to what it was created for.
At first, there was an idea to implement a service similar to how it is implemented in the MVC pattern.
But it turned out that it stores data and processes them.

2.3. MidExamConstants
Changed type from interface to regular class.
Constants are imported into other classes separately

2.4. SuperWindowProgram
The class was removed due to the lack of separation of one method into a separate class

2.5. SquareColumn
New class. Own implementation of LinkedList.
2.5.1. Nested class Node
This class stores the drawn square and a reference to the next and previous elements of the column
2.5.2. SquareColumnAnimator nested class
This class implements the Runnable interface, which allows you to run it in a separate thread

These two classes are both private and integral to the SquareColumn class. Therefore, it was decided to use nested classes.